+++ builder_pkgs/mockito/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.1.2
+
+* Produce a meaningful error message if an argument matcher is used outside of
+  stubbing (`when`) or verification (`verify` and `untilCalled`).
+
 ## 4.1.1
 
 * Mark the unexported and accidentally public `setDefaultResponse` as